Refactor custom fields schema selector
